chore(service-mesh): updates the lib folder names to match charmcraft (#533)
A minor update that updates the library path names of all libraries
owned by the service mesh team to match the interface name in
charmcraft.yaml
