--- interfaces/gateway-metadata/CHANGELOG.md
@@ -1,9 +1,0 @@
-# Changelog
-
-## 1.0.0.post0 - 16 April 2026
-
-Correct package metadata URLs.
-
-## 1.0.0 - 16 April 2026
-
-Initial release. Migrated from `charmed-service-mesh-helpers` (v0.6.0) `interfaces.gateway_metadata` module.
