--- interfaces/gateway-metadata/README.md
@@ -1,11 +1,0 @@
-# charmlibs.interfaces.gateway_metadata
-
-The `gateway-metadata` interface library.
-
-To install, add `charmlibs-interfaces-gateway-metadata` to your Python dependencies. Then in your Python code, import as:
-
-```py
-from charmlibs.interfaces import gateway_metadata
-```
-
-See the [reference documentation](https://documentation.ubuntu.com/charmlibs/reference/charmlibs/interfaces/gateway-metadata) for more.
