+++ interfaces/gateway_metadata/CHANGELOG.md
@@ -1,0 +1,13 @@
+# Changelog
+
+## 1.0.1 - 16 June 2026
+
+- Minor docstring and README updates.
+
+## 1.0.0.post0 - 16 April 2026
+
+Correct package metadata URLs.
+
+## 1.0.0 - 16 April 2026
+
+Initial release. Migrated from `charmed-service-mesh-helpers` (v0.6.0) `interfaces.gateway_metadata` module.
