+++ interfaces/gateway_metadata/src/charmlibs/interfaces/gateway_metadata/_version.py
