--- interfaces/istio-ingress-route/README.md
@@ -1,11 +1,0 @@
-# charmlibs.interfaces.istio_ingress_route
-
-The `istio-ingress-route` interface library.
-
-To install, add `charmlibs-interfaces-istio-ingress-route` to your Python dependencies. Then in your Python code, import as:
-
-```py
-from charmlibs.interfaces import istio_ingress_route
-```
-
-See the [reference documentation](https://documentation.ubuntu.com/charmlibs/reference/charmlibs/interfaces/istio-ingress-route) for more.
