--- interfaces/istio-request-auth/README.md
@@ -1,11 +1,0 @@
-# charmlibs.interfaces.istio_request_auth
-
-The `istio-request-auth` interface library.
-
-To install, add `charmlibs-interfaces-istio-request-auth` to your Python dependencies. Then in your Python code, import as:
-
-```py
-from charmlibs.interfaces import istio_request_auth
-```
-
-See the [reference documentation](https://documentation.ubuntu.com/charmlibs/reference/charmlibs/interfaces/istio-request-auth) for more.
