--- interfaces/istio-request-auth/pyproject.toml
@@ -1,74 +1,0 @@
-[project]
-name = "charmlibs-interfaces-istio-request-auth"
-description = "The charmlibs.interfaces.istio_request_auth package."
-readme = "README.md"
-requires-python = ">=3.10"
-authors = [
-    {name="Service Mesh"},
-]
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: Apache Software License",
-    "Intended Audience :: Developers",
-    "Operating System :: POSIX :: Linux",
-    "Development Status :: 5 - Production/Stable",
-]
-dynamic = ["version"]
-dependencies = [
-    "ops>=2.23.1,<4",
-    "pydantic>=2,<3",
-]
-
-[dependency-groups]
-lint = [  # installed for `just lint interfaces/istio-request-auth` (unit, functional, and integration are also installed)
-    # "typing_extensions",
-]
-unit = [  # installed for `just unit interfaces/istio-request-auth`
-    "ops[testing]",
-]
-functional = [  # installed for `just functional interfaces/istio-request-auth`
-]
-integration = [  # installed for `just integration interfaces/istio-request-auth`
-    "jubilant",
-]
-
-[project.urls]
-"Documentation" = "https://documentation.ubuntu.com/charmlibs/reference/charmlibs/interfaces/istio-request-auth"
-"Repository" = "https://github.com/canonical/charmlibs/tree/main/interfaces/istio-request-auth"
-"Issues" = "https://github.com/canonical/charmlibs/issues"
-"Changelog" = "https://github.com/canonical/charmlibs/blob/main/interfaces/istio-request-auth/CHANGELOG.md"
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.wheel]
-packages = ["src/charmlibs"]
-
-[tool.hatch.version]
-path = "src/charmlibs/interfaces/istio_request_auth/_version.py"
-
-[tool.ruff]
-extend = "../../pyproject.toml"
-src = ["src", "tests/unit", "tests/functional", "tests/integration"]  # correctly sort local imports in tests
-
-[tool.ruff.lint.extend-per-file-ignores]
-# add additional per-file-ignores here to avoid overriding repo-level config
-"tests/**/*" = [
-    # "E501",  # line too long
-]
-
-[tool.pyright]
-extends = "../../pyproject.toml"
-include = ["src", "tests"]
-pythonVersion = "3.10"  # check no python > 3.10 features are used
-
-[tool.charmlibs.functional]
-ubuntu = []  # ubuntu versions to run functional tests with, e.g. "24.04" (defaults to just "latest")
-pebble = []  # pebble versions to run functional tests with, e.g. "v1.0.0", "master" (defaults to no pebble versions)
-sudo = false  # whether to run functional tests with sudo (defaults to false)
-
-[tool.charmlibs.integration]
-# tags to run integration tests with (defaults to running once with no tag, i.e. tags = [''])
-# Available in CI in tests/integration/pack.sh and integration tests as CHARMLIBS_TAG
-tags = []  # Not used by the pack.sh and integration tests generated by the template
