+++ interfaces/istio_ingress_route/CHANGELOG.md
@@ -1,0 +1,14 @@
+# Changelog
+
+## 1.0.2 - 16 June 2026
+
+- Minor docstring and README updates.
+
+## 1.0.1 - 29 April 2026
+
+- pins ops to >=2.23.1,<4
+- pins pydantic to >=2,<3
+
+## 1.0.0
+
+Initial release. Migrated from `charms.istio_ingress_k8s.v0.istio_ingress_route` (v0.3).
