+++ interfaces/istio_request_auth/CHANGELOG.md
@@ -1,0 +1,21 @@
+# Changelog
+
+## 0.0.3 - 16 June 2026
+
+- Minor docstring and README updates.
+
+## 0.0.2 - 29 April 2026
+
+- pins ops to >=2.23.1,<4
+- pins pydantic to >=2,<3
+
+## 0.0.1 - 22 April 2026
+
+Initial release.
+
+The initial `istio-request-auth` interface library provides the following features:
+
+- Lets charms share their trusted JWT issuers and header-claim mappings as jwt_rules with the `istio-ingress-k8s` charms
+- Lets charms specify multiple trusted issuers and their corresponding header-claim mappings as a list of jwt_rules
+- Provides convienience data models for describing JWTRules
+- Models the top level data bag as datamodel
