--- interfaces/service-mesh/CHANGELOG.md
@@ -1,11 +1,0 @@
-# Changelog
-
-## 1.0.0
-
-Initial release. Migrated from `charms.istio_beacon_k8s.v0.service_mesh` (v0.19).
-
-## 1.0.1 - 29 April 2026
-
-- pins ops to >=2.23.1,<4
-- pins pydantic to >=2,<3
-- pins canonical-service-mesh to <1
