+++ interfaces/service_mesh/README.md
@@ -1,0 +1,11 @@
+# charmlibs.interfaces.service_mesh
+
+The `service_mesh` interface library.
+
+To install, add `charmlibs-interfaces-service-mesh` to your Python dependencies. Then in your Python code, import as:
+
+```py
+from charmlibs.interfaces import service_mesh
+```
+
+See the [reference documentation](https://documentation.ubuntu.com/charmlibs/reference/charmlibs/interfaces/service-mesh) for more.
